Download functionality working
